pyproject.toml
[project]
name = "django-recipe-repo"
readme = "README.md"
version = "0.1.0"
authors = [
    { name = "Jesse Sopel", email = "jesse.sopel@gmail.com" },
]
requires-python = ">=3.12"
dependencies = [
    "django",
    "django-stubs[compatible-mypy]",
    "django-admin-interface",
    "django-colorfield",
    "django-breeze",
    "django-environ",
    "django-cachalot",
    "django-treebeard",
    "django-modeltranslation",
    "easy-thumbnails",
    "pillow",
    "pint",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
  \.git
  | \.mypy_cache
  | \.ruff_cache
  | env
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.django-stubs]
django_settings_module = "myproject.settings"

[tool.ruff]
fix = true
line-length = 120
target-version = "py312"
extend-exclude = ["**migrations"]

[tool.ruff.lint]
ignore = ["D100", "D101", "D104", "D105", "D106", "D203", "D212", "PLR2004"]
select = [
    "A",
    "ASYNC",
    "B",
    "COM",
    "C4",
    "D",
    "E",
    "ERA",
    "F",
    "FA",
    "FLY",
    "G",
    "I",
    "N",
    "PTH",
    "PT",
    "PL",
    "PLE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SLF",
    "SIM",
    "TD",
    "TCH",
    "UP",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "D103",  # No docstrings in tests needed
    "S101",  # We needs asserts in tests
    "SLF001",
]
"**/migrations/*" = []
